Alexander's change. Fixes MTC bug by supporting 64bit processors in Timeval. Math was bad.
git-svn-id: http://wush.net/svn/range/software/public/openbts/trunk@2835 19bc5d8c-e614-43d4-8b26-e1612bc8e597

+++ CommonLibs/Timeval.cpp
@@ -71,8 +71,8 @@ double Timeval::seconds() const
 long Timeval::delta(const Timeval& other) const
 {
 	// 2^31 milliseconds is just over 4 years.
-	long deltaS = other.sec() - sec();
-	long deltaUs = other.usec() - usec();
+	int32_t deltaS = other.sec() - sec();
+	int32_t deltaUs = other.usec() - usec();
 	return 1000*deltaS + deltaUs/1000;
 }
